display embedded tweet on artefact page
